Nowy folder/vq_system/vq_system/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
jinja2
python-multipart
passlib
python-jose[cryptography]
pydantic
